requirements.txt
click>=8.0.0
docker>=6.0.0
rich>=13.0.0
PyYAML>=6.0.0
calimero-client-py>=0.1.4

# Development dependencies
black>=25.0.0
